Zip: correct portability issue with memory management
`free` should not be used to release the memory allocation.  The
allocation was performed with `UnsafeMutablePointer.allocate(capacity:)`
which does not provide a guarantee on the allocation being performed
with `malloc`.  In fact, on Windows, this is actively not the case and
is is performed with `_aligned_alloc`.  Therefore the release must be
balanced with the MSVC specific `_aligned_free`.  Instead, prefer to use
the portable `UnsafeMutablePointer.deallocate` to correctly balance the
memory allocation.