Zip: correct type conversion for LLP64 systems
The `unzGetCurrentFileInfo64` signature does not use a C type for the
length but rather a local type definition.  The type definition
underlying it is a `long` which is differently sized on LP64 and LLP64
systems.  Adjust this by casting the parameter to the proper type
(`uLong`) as vended by the C library.