minizip: correct preprocessor directive for portability
`#import` is an Objective-C construct and is not portable.  `#include`
is the standard mechanism for including a source file.  Use of the
Objective-C mechanism works in many cases but not globally.  On Windows,
this is used as a MSVC extension for COM type library inclusion.

Switch to the standard spelling to allow building on Windows.