minizip: remove the extraneous link directive
The linking is already handled at the SPM layer.  There is no reason to
force a secondary link directive.  This primarily impacts Darwin
platforms where the link directive would form a `LC_LINKER_OPTION` load
command to ensure that zlib is linked against.  However, we have the
platform specific linkage and this is already performed at the SPM
layer.  Remove the directive and allow SPM to handle it.