build: link against zlib properly on Windows
The library name for zlib is different on Windows than on Unix platforms
(`zlib` and `zlibstatic`).  In order to support properly linking against
the library the user must specify the linkage type.  Since SwiftPM does
not have the ability to properly control build settings, use an
environment variable to select between the two.  If
`ZIP_USE_DYNAMIC_ZLIB` is set, we will attempt to perform dynamic
linkage against zlib, otherwise, we will use static linking.

Take the opportunity to silence some warnings due to deprecated symbol
usage on Windows.

A build of zlib must be available for use and must be explicitly passed
to the build on Windows via extended flags (i.e. `-Xcc` and `-Xlinker`).